+++ db/pvault.sql
@@ -288,10 +288,6 @@ CREATE TABLE `pv_meta` (
 
 CREATE TABLE `settings` (
   `id` int(11) NOT NULL,
-  `label_printer_addr` varchar(255) COLLATE utf8_general_ci DEFAULT NULL,
-  `label_printer_model` varchar(225) COLLATE utf8_general_ci DEFAULT NULL,
-  `label_printer_size` varchar(255) COLLATE utf8_general_ci DEFAULT NULL,
-  `label_printer_font_size` int(11) DEFAULT 80,
   `currency` varchar(40) COLLATE utf8_general_ci DEFAULT NULL,
   `top_n` varchar(10) COLLATE utf8_general_ci NOT NULL,
   `heart_n` varchar(10) COLLATE utf8_general_ci NOT NULL,